requirements.txt
scenedetect[opencv]
ffmpeg-python
openai-whisper
langchain
langchain-openai
langchain-chroma
chromadb
streamlit
sentence-transformers
python-dotenv
Pillow
